In-memory Reminder module
Added a reminder module that operates in it's own thread, for the most part.
It's the first attempt and it's kinda scuffed to the python's multithreading limitations, thus involving a few workarounds and a check function in the main thread, though it does not seem to heavily impact performance currently.
Basic usage:  .remind 2020/01/26 21:23 This works, somewhat
format: .remind year/month/day hour:minute [and a message]
Accuracy +- 5 minutes.